pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "depthfusion"
version = "0.1.0"
description = "Depth-aware memory fusion for Claude Code — weighted retrieval, session attention, context routing, recursive LLMs"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
    "structlog>=24.0",
]

[project.optional-dependencies]
rlm = ["rlms"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.0",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]   # test assertions and names are often verbose
